% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burgle.R, R/burgle_CSC.R, R/burgle_cph.R,
%   R/burgle_flexsurv.R, R/burgle_lm.R, R/burgle_multinom.R, R/burgle_rfsrc.R,
%   R/burgle_survival.R
\name{burgle_}
\alias{burgle_}
\alias{burgle}
\alias{burgle.lm}
\alias{burgle.glm}
\alias{burgle.biglm}
\alias{burgle.bigglm}
\alias{burgle.speedlm}
\alias{burgle.speedglm}
\alias{burgle.fastLm}
\alias{burgle.gam}
\alias{burgle.CauseSpecificCox}
\alias{burgle.cph}
\alias{burgle.flexsurvreg}
\alias{burgle.multinom}
\alias{burgle.rfsrc}
\alias{burgle.coxph}
\title{Burgle}
\usage{
burgle(object, ...)

\method{burgle}{lm}(object, ...)

\method{burgle}{glm}(object, ...)

\method{burgle}{biglm}(object, ...)

\method{burgle}{bigglm}(object, ...)

\method{burgle}{speedlm}(object, ...)

\method{burgle}{speedglm}(object, ...)

\method{burgle}{fastLm}(object, ...)

\method{burgle}{gam}(object, ...)

\method{burgle}{CauseSpecificCox}(object, ...)

\method{burgle}{cph}(object, ...)

\method{burgle}{flexsurvreg}(object, ...)

\method{burgle}{lm}(object, ...)

\method{burgle}{multinom}(object, ...)

\method{burgle}{rfsrc}(object, ...)

\method{burgle}{coxph}(object, ...)
}
\arguments{
\item{object}{the model object to burgle}

\item{...}{must be left empty for now}
}
\value{
a burgle_ object
}
\description{
Burgling what is necessary from different objects
}
\examples{
fit <- lm(Sepal.Length ~ Sepal.Width + Petal.Length, data = iris)
bfit <- burgle(fit)
object.size(fit)
object.size(bfit)
}
