% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_time.R
\name{predict_time}
\alias{predict_time}
\title{Title}
\usage{
predict_time(object, ...)
}
\arguments{
\item{object}{a burgle object, used in with survival time}

\item{...}{}
}
\value{
a dataframe of with a smple
}
\description{
Title
}
\examples{
library(survival)
lung <- survival::lung
cx <- burgle(coxph(Surv(time, status-1)~age + sex, data = lung))
predict_time(cx, newdata = head(lung))
}
