% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burgle_flexsurv.R, R/burgle_survival.R,
%   R/predict_burgle.R, R/simulate_models.R
\name{simulate_models}
\alias{simulate_models}
\alias{simulate_models.burgle_flexsurvreg}
\alias{simulate_models.burgle_coxph}
\alias{simulate_models.burgle_lm}
\alias{simulate_models.burgle_glm}
\title{Simulate Models}
\usage{
\method{simulate_models}{burgle_flexsurvreg}(
  object,
  models = NULL,
  newdata = NA,
  type = "lp",
  sims = 1,
  seed = NULL,
  times = NULL,
  ...
)

\method{simulate_models}{burgle_coxph}(
  object,
  models,
  newdata,
  type = "lp",
  sims = 1,
  seed = NULL,
  times = NULL,
  ...
)

\method{simulate_models}{burgle_lm}(
  object,
  models = NULL,
  newdata,
  type = "lp",
  sims = 1,
  seed = NULL,
  se = FALSE,
  limits = NULL,
  se_type = "prediction",
  ...
)

\method{simulate_models}{burgle_glm}(
  object,
  models = NULL,
  newdata,
  type = "lp",
  sims = 1,
  seed = NULL,
  se = FALSE,
  se_type = "prediction",
  ...
)

simulate_models(
  object,
  models,
  newdata,
  type = "lp",
  sims = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{object}{the results of burgle_* object}

\item{models}{a matrix or vector of model coefficeints, one can specifically it or get from `draw_models()`}

\item{newdata}{new data of class data.frame}

\item{type}{either 'lp', 'response', 'link' for glm or 'risk' if time dependent}

\item{sims}{how many simulated response to draw}

\item{seed}{seed to set for reproducibility}

\item{times}{if type = "risk" time for which to predict risk, if times and sims is multiple the return will be lists within lists}

\item{...}{arguemnts for specific methods}

\item{se}{whether or not to include the standard error in the simulations}

\item{limits}{limits (minimum and maximum) for simulated response values}

\item{se_type}{either 'prediction' or 'confidence' for standard errors use in simulating}
}
\description{
Simulating from burgle_ objects and draw_models()
}
