% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burgle.R, R/burgle_CSC.R, R/burgle_flexsurv.R,
%   R/burgle_multinom.R, R/burgle_survival.R
\name{burgle_}
\alias{burgle_}
\alias{burgle}
\alias{burgle.lm}
\alias{burgle.glm}
\alias{burgle.CauseSpecificCox}
\alias{burgle.flexsurvreg}
\alias{burgle.multinom}
\alias{burgle.coxph}
\title{Burgle}
\usage{
burgle(object, ...)

\method{burgle}{lm}(object, ...)

\method{burgle}{glm}(object, ...)

\method{burgle}{CauseSpecificCox}(object, ...)

\method{burgle}{flexsurvreg}(object, ...)

\method{burgle}{multinom}(object, ...)

\method{burgle}{coxph}(object, ...)
}
\arguments{
\item{object}{the model object to burgle}

\item{...}{must be left empty for now}
}
\value{
a burgle_ object
}
\description{
Burgling what is necessary from different objects
}
\examples{
fit <- lm(Sepal.Length ~ Sepal.Width + Petal.Length, data = iris)
bfit <- burgle(fit)
object.size(fit)
object.size(bfit)
}
